requirements.txt
absl-py==0.9.0
apache-beam==2.23.0
appnope==0.1.0
argon2-cffi==20.1.0
astunparse==1.6.3
attrs==19.3.0
avro-python3==1.9.2.1
backcall==0.2.0
bleach==3.1.5
boto3==1.14.37
botocore==1.17.37
cachetools==3.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
crcmod==1.7
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
dm-tree==0.1.5
docker==4.2.2
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
fastavro==0.23.6
fasteners==0.15
Flask==1.1.2
future==0.18.2
gast==0.3.3
google-api-core==1.22.0
google-api-python-client==1.10.0
google-apitools==0.5.31
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.24.0
google-cloud-bigtable==1.0.0
google-cloud-core==1.4.1
google-cloud-datastore==1.7.4
google-cloud-dlp==0.13.0
google-cloud-language==1.3.0
google-cloud-pubsub==1.0.2
google-cloud-spanner==1.13.0
google-cloud-videointelligence==1.13.0
google-cloud-vision==0.42.0
google-pasta==0.2.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.31.0
grpcio-gcp==0.2.2
h5py==2.10.0
hdfs==2.5.8
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
keras-tuner==1.0.1
kiwisolver==1.2.0
kubernetes==11.0.0
logzero==1.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
ml-metadata==0.22.1
mock==2.0.0
monotonic==1.5
more-itertools==8.4.0
mpmath==1.1.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.1
numpy==1.19.1
oauth2client==4.1.3
oauthlib==3.1.0
opencv-contrib-python==4.3.0.36
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.4
ptyprocess==0.6.0
py==1.9.0
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.6.1
pymongo==3.11.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
pytest-flakes==4.0.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-fairness-indicators==0.0.5
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-data-validation==0.22.2
tensorflow-estimator==2.2.0
tensorflow-hub==0.8.0
tensorflow-metadata==0.22.2
tensorflow-model-analysis==0.22.2
tensorflow-privacy==0.3.0
tensorflow-serving-api==2.2.0
tensorflow-transform==0.22.0
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
tfx==0.22.0
tfx-bsl==0.22.1
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
witwidget==1.7.0
wrapt==1.12.1
zipp==3.1.0
